// frontend/src/app/core/models/message.model.ts
// message.model.ts
export interface Message {
    id: number;
    messageId: string;
    queueName: string;
    content: string;
    messageType: string;
    receivedAt: string;
    processed: boolean;
    processedAt?: string;
  }
  
  export interface Page<T> {
    content: T[];
    pageable: {
      pageNumber: number;
      pageSize: number;
      sort: {
        sorted: boolean;
        unsorted: boolean;
        empty: boolean;
      };
      offset: number;
      paged: boolean;
      unpaged: boolean;
    };
    totalElements: number;
    totalPages: number;
    last: boolean;
    size: number;
    number: number;
    sort: {
      sorted: boolean;
      unsorted: boolean;
      empty: boolean;
    };
    numberOfElements: number;
    first: boolean;
    empty: boolean;
  }